requirements.txt
absl-py==0.15.0
addict==2.4.0
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
basicsr==1.3.4.7
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
bleach==4.1.0
brotlipy==0.7.0
cachetools==4.2.4
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
conda==4.9.2
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
cycler==0.10.0
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
dlib==19.22.1
dnspython==2.1.0
dominate==2.6.0
entrypoints==0.3
face-alignment==1.3.5
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
future==0.18.2
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
google-auth==2.3.2
google-auth-oauthlib==0.4.6
grpcio==1.41.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.9.0
imageio-ffmpeg==0.4.5
ipykernel==6.4.2
ipython @ file:///tmp/build/80754af9/ipython_1610724799192/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.5
jedi @ file:///tmp/build/80754af9/jedi_1606932564285/work
Jinja2 @ file:///home/linux1/recipes/ci/jinja2_1610990516718/work
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
kornia==0.5.5
libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
llvmlite==0.37.0
lmdb==1.2.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.5
numba==0.54.1
numpy==1.19.0
nvdiffrast @ file:///apdcephfs/share_1290939/feiiyin/TH/PIRender/Deep3DFaceRecon_pytorch/nvdiffrast
oauthlib==3.1.1
olefile==0.46
opencv-python==3.4.9.33
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.7.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==6.2.1
pkginfo==1.7.0
prometheus-client==0.11.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.19.0
psutil @ file:///tmp/build/80754af9/psutil_1598370257551/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1610565767015/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==3.0.1
pyrsistent==0.18.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-etcd==0.4.5
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel-yaml==0.15.87
scikit-image==0.16.2
scipy==1.4.1
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1605205327372/work
soupsieve @ file:///tmp/build/80754af9/soupsieve_1607965878077/work
tb-nightly==2.8.0a20211106
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.12.1
testpath==0.5.0
torch==1.7.1
torchelastic==0.2.1
torchvision==0.8.2
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
trimesh==3.9.20
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
yapf==0.31.0
